feat(UIKIT-972): Правки по PR
