+++ packages/components/src/theme/palette/componentsColors/componentsColors.ts
@@ -1,0 +1,7 @@
+import { linkColors } from './linkColors';
+
+export type ComponentsColors = typeof componentsColors;
+
+export const componentsColors = {
+  link: linkColors,
+};
