+++ packages/components/src/theme/palette/componentsColors/linkColors/index.ts
@@ -1,6 +1,1 @@
-export const linkColors = {
-  main: '#1874FF',
-  hover: '#268FF9',
-  visited: '#762376',
-  active: '#072D57',
-};
+export * from './linkColors';
