+++ packages/components/src/theme/palette/componentsColors/linkColors/linkColors.ts
@@ -1,0 +1,6 @@
+export const linkColors = {
+  main: '#1874FF',
+  hover: '#268FF9',
+  visited: '#762376',
+  active: '#072D57',
+};
